Move kwerdenker's personal keymap from RGB (#7645)
* Move kwerdenker's personal keymap from RGB

Talked to Maartenwut about making RGB an actual template instead of
kwerdenker's keymap. Free'd up B7 so it can be used for audio in
agreement with MW. He's considering adding a speaker footprint to next
rev of plain60, and only B7 can be used with Audio. RGB uses bitbang
driver so it doesn't really make any difference if it's pin B0.

 - RGB should be a template for other people to modify, not a personal
 keymap
 - change default pin to B0 to free up B7 for audio use
 - rename RGB to kwerdenker

* Add RGB keycodes to FN layer

* don't use uppercase in keymap names

 - doesn't work properly on MacOS or Windows

* Add an audio specific keymap

* Enum instead of define

* Suggested change to keymap

 - maintain compatibility with old keymap